// types.ts
export interface Service {
    id: number;
    clientName: string;
    location: string;
    serviceDate: string;
    durationHours: number;
    kilometers: number;
    description: string;
    agreedPrice?: number;
    isSettled: boolean;
    createdBy: string;
}